Ajuste variavel tela de login
