+++ src/com/company/Main.java
@@ -1,6 +1,6 @@
 package com.company;
 
-import com.company.controllers.CorbaViewController;
+import com.company.controllers.CorbaLoginViewController;
 import javafx.application.Application;
 import javafx.stage.Stage;
 
@@ -8,8 +8,8 @@ public class Main extends Application {
 
     @Override
     public void start(Stage primaryStage) throws Exception {
-        CorbaViewController corbaViewController = new CorbaViewController(primaryStage);
-        corbaViewController.create();
+        CorbaLoginViewController corbaLoginViewController = new CorbaLoginViewController(primaryStage);
+        corbaLoginViewController.create();
     }
     public static void main(String args[]){
         launch(args);
