+++ src/com/company/controllers/CorbaLoginViewController.java
@@ -1,0 +1,24 @@
+package com.company.controllers;
+
+import com.company.views.CorbaLoginView;
+import javafx.stage.Stage;
+
+public class CorbaLoginViewController {
+
+    private Stage stage;
+    private CorbaLoginView corbaLoginView;
+
+    public CorbaLoginViewController(Stage stage){
+        this.corbaLoginView = new CorbaLoginView();
+        this.stage = stage;
+    }
+
+    public void create(){
+        this.corbaLoginView.createScene();
+        this.corbaLoginView.createCorbaViewScene();
+
+        this.stage.setTitle("Corba File Manager");
+        this.stage.setScene(this.corbaLoginView.getScene());
+        this.stage.show();
+    }
+}
